Add test demo and form code
